Start work on resolving known problems with tables
Sometimes the source data includes quoting errors and other things that
we know about (because these are historical files) and that we know how
to deal with.

Here, we create a place for those changes to be implemented (but don't
actually implement any of them yet), and we continue to pass warnings on
to the user when a particular problem hasn't yet been addressed.

One caveat, though, is that in order to format the problems in a
familiar way, I used most of the implementation from readr, which is
licensed under GPL version 2 or later. That means this package also
should be subject to the same license.